LegalDocumentChatBot/requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.30.0
langchain>=0.0.300
langchain-community>=0.0.20
langchain-huggingface>=0.0.3
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
accelerate>=0.20.0